Добавляет презентер для карточки фильма
